// src/generated/models/Hx_tasksesModel.ts
export const Hx_tasksescr18c_tasksource = {
  989230000: 'Teams',
  989230001: 'TMS',
  989230002: 'Leadership',
  989230003: 'Steering',
  989230004: 'Planning&Monitoring',
  989230005: 'Strategy',
  989230007: 'Projects',
  989230008: 'Objectives',
  989230009: 'Excel'
} as const;
export type Hx_tasksescr18c_tasksource = keyof typeof Hx_tasksescr18c_tasksource;
export const Hx_taskseshx_businessunit = {
  123200000: 'BU1',
  123200001: 'BU2',
  123200002: 'BU3'
} as const;
export type Hx_taskseshx_businessunit = keyof typeof Hx_taskseshx_businessunit;
export const Hx_taskseshx_delaynotified = {
  123200000: 'True',
  123200001: 'False'
} as const;
export type Hx_taskseshx_delaynotified = keyof typeof Hx_taskseshx_delaynotified;
export const Hx_taskseshx_extendduedate = {
  123200000: 'Yes',
  123200001: 'No'
} as const;
export type Hx_taskseshx_extendduedate = keyof typeof Hx_taskseshx_extendduedate;
export const Hx_taskseshx_extendduedatestatus = {
  123200000: 'Approved',
  123200001: 'Rejected',
  123200002: 'Pending'
} as const;
export type Hx_taskseshx_extendduedatestatus = keyof typeof Hx_taskseshx_extendduedatestatus;
export const Hx_taskseshx_holdrequeststatus = {
  123200000: 'Pending',
  123200001: 'Approved ',
  123200002: 'Rejected '
} as const;
export type Hx_taskseshx_holdrequeststatus = keyof typeof Hx_taskseshx_holdrequeststatus;
export const Hx_taskseshx_level = {
  123200000: 'Updates',
  123200001: 'Milestone',
  123200002: 'Corrective Action'
} as const;
export type Hx_taskseshx_level = keyof typeof Hx_taskseshx_level;
export const Hx_taskseshx_priority = {
  123200000: 'Low',
  123200001: 'Medium',
  123200002: 'High',
  931940001: 'Critical'
} as const;
export type Hx_taskseshx_priority = keyof typeof Hx_taskseshx_priority;
export const Hx_taskseshx_recurrenceschedule = {
  123200000: 'Daily',
  123200001: 'Weekly',
  123200002: 'Monthly',
  123200003: 'Yearly'
} as const;
export type Hx_taskseshx_recurrenceschedule = keyof typeof Hx_taskseshx_recurrenceschedule;
export const Hx_taskseshx_recurrencetype = {
  123200000: 'One-Time',
  123200001: 'Recurring'
} as const;
export type Hx_taskseshx_recurrencetype = keyof typeof Hx_taskseshx_recurrencetype;
export const Hx_taskseshx_starttask = {
  123200000: 'Yes',
  123200001: 'No'
} as const;
export type Hx_taskseshx_starttask = keyof typeof Hx_taskseshx_starttask;
export const Hx_taskseshx_status = {
  100000001: 'In Progress',
  100000005: 'On Hold',
  123200002: 'Closed',
  123200003: 'Cancelled',
  123200004: 'New',
  123200005: 'Submitted',
  931940001: 'Rejected'
} as const;
export type Hx_taskseshx_status = keyof typeof Hx_taskseshx_status;
export const Hx_taskseshx_submissionstatus = {
  123200001: 'Pending Review',
  123200002: 'Approved',
  123200003: 'Rejected'
} as const;
export type Hx_taskseshx_submissionstatus = keyof typeof Hx_taskseshx_submissionstatus;
export const Hx_taskseshx_taskonhold = {
  123200000: 'Yes',
  123200001: 'No'
} as const;
export type Hx_taskseshx_taskonhold = keyof typeof Hx_taskseshx_taskonhold;
export const Hx_taskseshx_taskvisibility = {
  123200000: 'Department',
  123200002: 'Single'
} as const;
export type Hx_taskseshx_taskvisibility = keyof typeof Hx_taskseshx_taskvisibility;
export const Hx_tasksesobjectiv_functioncontribution = {
  1: 'Major',
  2: 'Intermediate',
  3: 'Minor'
} as const;
export type Hx_tasksesobjectiv_functioncontribution = keyof typeof Hx_tasksesobjectiv_functioncontribution;
export const Hx_tasksesobjectiv_kpiimpact = {
  1: 'Financial',
  2: 'Non Financial'
} as const;
export type Hx_tasksesobjectiv_kpiimpact = keyof typeof Hx_tasksesobjectiv_kpiimpact;
export const Hx_tasksesobjectiv_taskkpitype = {
  1: 'Positive',
  2: 'Negative'
} as const;
export type Hx_tasksesobjectiv_taskkpitype = keyof typeof Hx_tasksesobjectiv_taskkpitype;
export const Hx_tasksesobjectiv_tasklevel = {
  1: 'Process',
  2: 'Project'
} as const;
export type Hx_tasksesobjectiv_tasklevel = keyof typeof Hx_tasksesobjectiv_tasklevel;
export const Hx_tasksesobjectiv_taskownerlevel = {
  1: 'Manager',
  2: 'Supervisor',
  3: 'Head',
  4: 'Senior',
  5: 'Specialist'
} as const;
export type Hx_tasksesobjectiv_taskownerlevel = keyof typeof Hx_tasksesobjectiv_taskownerlevel;
export const Hx_tasksesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Hx_tasksesstatecode = keyof typeof Hx_tasksesstatecode;
export const Hx_tasksesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Hx_tasksesstatuscode = keyof typeof Hx_tasksesstatuscode;
export const Hx_tasksestms_escalate = {
  1: 'Yes',
  0: 'No'
} as const;
export type Hx_tasksestms_escalate = keyof typeof Hx_tasksestms_escalate;
export const Hx_tasksestms_escalationoutcome = {
  931940000: 'Resolved',
  931940001: 'Withdrawn',
  931940002: 'Closed Unresolved',
  931940003: 'Void'
} as const;
export type Hx_tasksestms_escalationoutcome = keyof typeof Hx_tasksestms_escalationoutcome;
export const Hx_tasksestms_escalationresolutionmethod = {
  931940000: 'Reassigned',
  931940001: 'Escalation-Related Subtask Created',
  931940002: 'Comment Returned to Requestor'
} as const;
export type Hx_tasksestms_escalationresolutionmethod = keyof typeof Hx_tasksestms_escalationresolutionmethod;
export const Hx_tasksestms_reassign = {
  1: 'Yes',
  0: 'No'
} as const;
export type Hx_tasksestms_reassign = keyof typeof Hx_tasksestms_reassign;
export const Hx_tasksestms_resubmittask = {
  1: 'Yes',
  0: 'No'
} as const;
export type Hx_tasksestms_resubmittask = keyof typeof Hx_tasksestms_resubmittask;
export const Hx_tasksestms_submittask = {
  1: 'Yes',
  0: 'No'
} as const;
export type Hx_tasksestms_submittask = keyof typeof Hx_tasksestms_submittask;
export const Hx_tasksestms_taskstatuscalc = {
  931940000: 'New',
  931940001: 'In Progress',
  931940002: 'Submitted',
  931940003: 'Closed',
  931940004: 'Rejected',
  931940005: 'On Hold',
  931940006: 'Cancelled'
} as const;
export type Hx_tasksestms_taskstatuscalc = keyof typeof Hx_tasksestms_taskstatuscalc;

export interface Hx_tasksesBase {
  cr18c_id?: string;
  "cr18c_KPITask@odata.bind"?: string;
  "cr18c_RelatedLeadershipPractice@odata.bind"?: string;
  "cr18c_Specialty@odata.bind"?: string;
  cr18c_spitemid?: string;
  cr18c_tasksource: Hx_tasksescr18c_tasksource;
  cr603_posturl?: string;
  "hx_Accountable@odata.bind"?: string;
  hx_actualstartdate?: string;
  hx_approvaldate?: string;
  "hx_Assignee@odata.bind": string;
  hx_assigneenotificationsent?: boolean;
  hx_attachlink?: string;
  hx_businessunit?: Hx_taskseshx_businessunit;
  hx_completiondate?: string;
  hx_consulted?: string;
  hx_delaynotified?: Hx_taskseshx_delaynotified;
  "hx_DependentTask@odata.bind"?: string;
  hx_duedate: string;
  hx_extendduedate?: Hx_taskseshx_extendduedate;
  hx_extendduedatereason?: string;
  hx_extendduedatestatus?: Hx_taskseshx_extendduedatestatus;
  hx_holdrequestreason?: string;
  hx_holdrequeststatus?: Hx_taskseshx_holdrequeststatus;
  "hx_Informed@odata.bind"?: string;
  hx_justifications?: string;
  hx_level?: Hx_taskseshx_level;
  "hx_ManagerOne@odata.bind"?: string;
  "hx_ManagerTwo@odata.bind"?: string;
  "hx_Organization@odata.bind"?: string;
  hx_priority?: Hx_taskseshx_priority;
  hx_progress?: number;
  "hx_RaisedBy@odata.bind": string;
  hx_recurrenceenddate?: string;
  hx_recurrenceschedule?: Hx_taskseshx_recurrenceschedule;
  hx_recurrencetype?: Hx_taskseshx_recurrencetype;
  "hx_RelatedMilestone@odata.bind"?: string;
  "hx_RelatedProject@odata.bind"?: string;
  hx_responsible?: string;
  "hx_Source@odata.bind"?: string;
  hx_startdate: string;
  hx_starttask?: Hx_taskseshx_starttask;
  hx_status?: Hx_taskseshx_status;
  hx_submissionstatus?: Hx_taskseshx_submissionstatus;
  "hx_TargetedDepartment_1@odata.bind"?: string;
  hx_taskcode?: string;
  hx_taskdescription?: string;
  hx_taskonhold?: Hx_taskseshx_taskonhold;
  "hx_TaskParent@odata.bind"?: string;
  hx_tasksid: string;
  hx_tasktitle: string;
  hx_taskvisibility?: Hx_taskseshx_taskvisibility;
  hx_testing?: string;
  hx_userenddate?: string;
  hx_userstartdate?: string;
  importsequencenumber?: number;
  objectiv_achievement?: number;
  objectiv_actual?: number;
  objectiv_actualdate?: string;
  "objectiv_AssignTo@odata.bind"?: string;
  objectiv_functioncontribution?: Hx_tasksesobjectiv_functioncontribution;
  objectiv_growthachievement?: number;
  objectiv_kpiimpact?: Hx_tasksesobjectiv_kpiimpact;
  objectiv_linkofevidence?: string;
  "objectiv_MainDepartmentKPI@odata.bind"?: string;
  "objectiv_MainObjective@odata.bind"?: string;
  "objectiv_ObjectiveWeight@odata.bind"?: string;
  objectiv_percentageweight?: number;
  "objectiv_Process@odata.bind"?: string;
  objectiv_score?: number;
  "objectiv_SubProcess@odata.bind"?: string;
  objectiv_target?: number;
  objectiv_taskbaseline?: number;
  objectiv_taskkpitype?: Hx_tasksesobjectiv_taskkpitype;
  objectiv_tasklevel?: Hx_tasksesobjectiv_tasklevel;
  "objectiv_TaskOwner@odata.bind"?: string;
  objectiv_taskownerlevel?: Hx_tasksesobjectiv_taskownerlevel;
  objectiv_weight?: number;
  overriddencreatedon?: string;
  "pm_KPIAchievement@odata.bind"?: string;
  processid?: string;
  project_actualenddate?: string;
  "project_FollowUp@odata.bind"?: string;
  "project_Update@odata.bind"?: string;
  stageid?: string;
  statecode: Hx_tasksesstatecode;
  statuscode?: Hx_tasksesstatuscode;
  "stf_SourcePOC@odata.bind"?: string;
  "stf_SourceTactic@odata.bind"?: string;
  timezoneruleversionnumber?: number;
  "tms_DirectManagerLkup@odata.bind": string;
  tms_escalate?: Hx_tasksestms_escalate;
  "tms_EscalateBy@odata.bind"?: string;
  tms_escalateclearedon?: string;
  tms_escalateon?: string;
  "tms_EscalateTo@odata.bind"?: string;
  "tms_EscalationClearedBy@odata.bind"?: string;
  tms_escalationoutcome?: Hx_tasksestms_escalationoutcome;
  tms_escalationreason?: string;
  tms_escalationrelated?: boolean;
  tms_escalationreminder?: boolean;
  tms_escalationresolutioncomment?: string;
  tms_escalationresolutionmethod?: Hx_tasksestms_escalationresolutionmethod;
  tms_excelid?: string;
  tms_extendnewduedate?: string;
  "tms_OriginalAssignee@odata.bind"?: string;
  tms_reassign?: Hx_tasksestms_reassign;
  tms_rejectionreason?: string;
  tms_reminderuntil?: string;
  tms_resubmittask?: Hx_tasksestms_resubmittask;
  tms_submittask?: Hx_tasksestms_submittask;
  "tms_TaskCreator@odata.bind"?: string;
  traversedpath?: string;
  utcconversiontimezonecode?: number;
}